+++ Project/MapNoReduce/Common/RemoteInterfaces/IWorker.cs
@@ -5,8 +5,8 @@ namespace PADIMapNoReduce
 {
     public interface IWorker
     {
-       void ReceiveWork(string clientURL, long fileSize, int splits, string mapperName, byte[] mapperCode);
-       void FetchWorker(string clientURL, string jobTrackerURL, string mapperName, byte[] mapperCode, long fileSize, int totalSplits, int remainingSplits);
+       void ReceiveWork(string clientURL, long fileSize, long splits, string mapperName, byte[] mapperCode);
+       void FetchWorker(string clientURL, string jobTrackerURL, string mapperName, byte[] mapperCode, long fileSize, long totalSplits, long remainingSplits);
        bool IsAlive();
        List<string> AddWorker(string entryURL, bool firstContact);
